refactor: 축제 관리자 신청 상태 카드 분리
